requirements.txt
streamlit==1.29.0
opencv-python-headless==4.8.1.78
numpy==1.26.0
Pillow==10.0.1